finished moving support files to lib/src
