+++ lib/src/error_group.dart
